Move serialize/deserialize methods near to the usage
